Remove the console-logging feature
This feature has caused some confusion in the past, and is non-standard.
It's also not really that necessary, if users have to manually enable it
_anyway_. Instead, for our own code, we just default to using
`env_logger` directly. For any users this means changing their
`RUST_OPCUA_LOG=` to `RUST_LOG=opcua=`, or something like that. Most
libraries leave this to the user to do.