Add tracing to the server
This adds layered tracing to the server, and cleans up what we log a bit
(logging the message chunks is insane).

Tracing is nice and automatic, though you do have to be careful about
tracing over await points. This should make it possible (with the
correct subscriber) to produce structured spans for each request on the
server. When not using tracing, it provides decent debug-level logs with
some level of detail.

We trace connection establishment, requests on the top level, as well as
per call to a node manager. Node managers can do their own tracing
and logging, which will automatically be part of the top level span.