Do sequence numbers correctly
Currently we lack logic for rollover when we run out of sequence
headers, and also we lack the logic to deal with "legacy" and
"non-legacy" sequence numbers.

Why there is a difference is beyond me, there REALLY does not need to
be, but that's how it is.

This does not actually fully implement the non-legacy sequence numbers,
since the server needs to update its sequence numbers with the new
security policy (the client picks security policy before it creates the
channel). Since we don't (yet) support ECC at all, it doesn't affect us,
but we may need to handle that correctly in the future.