Fix panic in monitored item
This happens if a value is written that is _earlier_ than the last
reported value. In this case the server is saying that an old value is
newer than the newest value, which doesn't really make much sense, but
it would create difficulties if we didn't accept it. We just treat it as
a new value. Node managers that want to prevent users from writing
values in the past should do so.