Pause publish loop when receiving BadNoSubscription.
Reading the standard a bit more, it seems like
`BadSubscriptionIdInvalid` isn't a valid service-level status code for
`Publish`, so we don't need to consider that one here. As for
`BadNoSubscription`, if this happens we are _probably_ in the process of
creating a new subscription, so we can pause the automatic publishing
until it is created. For safety we also unpause if we successfully
receive a publish response.