Make subscription services stateless (#110)
This detaches the session subscription state from the direct service
implementations, making them retryable, and also making it possible to
use them without relying on the session state. It also adds `Publish`
and `Republish` as public services.

This further consolidates the session state, which I hope can make it
easier to completely detach and isolate the state from the session.

The breakage is minimal, only if you were relying on the subscription
direct service calls.

I also added a test for Republish, which is now a lot more possible to
do.